requirements.txt
playwright>=1.40.0
requests>=2.28.0
pyyaml>=6.0
